backend/requirements.txt
# Core
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0
djangorestframework-simplejwt>=5.3.0,<6.0.0
python-dotenv>=1.0.0,<2.0.0

# Database
psycopg2-binary>=2.9.9,<3.0.0

# CORS
django-cors-headers>=4.3.0,<5.0.0

# API Documentation
drf-yasg>=1.21.0,<2.0.0

# Debugging
django-debug-toolbar>=4.2.0,<5.0.0

# File Upload
Pillow>=10.0.0,<11.0.0

# Data Processing
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0

# DICOM
pydicom>=2.4.0,<3.0.0

# Email
python-decouple>=3.8,<4.0.0
django-environ>=0.11.0,<1.0.0

# Testing
pytest>=7.4.0,<8.0.0
pytest-django>=4.7.0,<5.0.0
pytest-cov>=4.1.0,<5.0.0

# Code Quality
black>=23.0.0,<24.0.0
flake8>=6.0.0,<7.0.0
isort>=5.0.0,<6.0.0

# Security
python-jose>=3.3.0,<4.0.0
cryptography>=41.0.0,<42.0.0

# Async
channels>=4.0.0,<5.0.0
channels-redis>=4.1.0,<5.0.0

# Other
django-filter>=23.0,<24.0
whitenoise>=6.0.0,<7.0.0
gunicorn>=21.0.0,<22.0.0
